refactor(orchestra): 公共接口统一 + 消除内部导入 (46 violations fixed)
修复 orchestra 模块内所有违规导入，统一使用公共 API：

**核心修改**：
1. orchestra/__init__.py
   - 添加 MAX_INTENTS_PER_TICK 到公共 API（重新导出）
   - 修复所有深层导入为公共 API 导入（clients, models, observability, services）
   - 修复 append_governance_event 导入路径

2. orchestra 子模块（9 个文件）
   - dispatch_coordinator_factory.py (13 fixes)
   - queue_operations.py (13 fixes)
   - queue_persistence_service.py (8 fixes)
   - issue_loader.py (4 fixes)
   - dispatch_health_check.py (3 fixes)
   - protocols.py (2 fixes)
   - logging.py (1 fix)
   - queue_entry.py (1 fix)
   - global_dispatch_coordinator.py (1 fix)

**验收标准**：
✅ orchestra 模块违规导入 = 0
✅ orchestra/__init__.py 有明确的 __all__ 列表
✅ 所有测试通过（134 passed）
✅ 类型检查通过（mypy strict）
✅ Lint 检查通过（ruff）

**合并冲突解决**：
- 合并 origin/main 时解决 orchestra/__init__.py 冲突
- 新增符号：FailedGate, append_governance_event（来自上游）
- 统一应用公共 API 导入模式

关闭 #2135
