+++ src/vibe3/orchestra/global_dispatch_coordinator.py
@@ -9,9 +9,9 @@ All symbols are also available from their canonical locations:
 
 # Re-export GlobalDispatchCoordinator from domain
 from vibe3.domain import GlobalDispatchCoordinator
-from vibe3.domain.dispatch_coordinator import MAX_INTENTS_PER_TICK
 
+# Re-export MAX_INTENTS_PER_TICK from orchestra (re-exported from domain)
 # Re-export QueueEntry from orchestra (still in orchestra layer)
-from vibe3.orchestra import QueueEntry
+from vibe3.orchestra import MAX_INTENTS_PER_TICK, QueueEntry
 
 __all__ = ["GlobalDispatchCoordinator", "QueueEntry", "MAX_INTENTS_PER_TICK"]
