+++ src/vibe3/orchestra/queue_entry.py
@@ -5,4 +5,4 @@
     This module is retained as a compatibility shell.
 """
 
-from vibe3.models.queue_entry import QueueEntry  # noqa: F401 — compatibility re-export
+from vibe3.models import QueueEntry  # noqa: F401 — compatibility re-export
