+++ src/vibe3/orchestra/queue_operations.py
@@ -4,26 +4,22 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable
 
-from vibe3.domain.qualify_gate import QualifyGateService
-from vibe3.domain.role_resolver import find_role_for_state
-from vibe3.models import IssueInfo, IssueState
-from vibe3.models.orchestra_config import OrchestraConfig
-from vibe3.models.queue_entry import QueueEntry
-from vibe3.orchestra.issue_loader import (
+from vibe3.models import IssueInfo, IssueState, OrchestraConfig, QueueEntry
+from vibe3.orchestra import (
+    QualifyGateService,
+    append_orchestra_event,
+    find_role_for_state,
     get_flow_context,
     is_auto_task_branch,
     load_issue,
+    sort_ready_issues,
 )
-from vibe3.orchestra.logging import append_orchestra_event
-from vibe3.orchestra.queue_ordering import sort_ready_issues
-from vibe3.services.label_utils import should_skip_from_queue
-from vibe3.services.orchestra_helpers import get_manager_usernames
+from vibe3.services import get_manager_usernames, should_skip_from_queue
 
 if TYPE_CHECKING:
-    from vibe3.clients.github_client import GitHubClient
-    from vibe3.clients.sqlite_client import SQLiteClient
-    from vibe3.environment.session_registry import SessionRegistryService
-    from vibe3.orchestra.protocols import FlowManagerProtocol
+    from vibe3.clients import GitHubClient, SQLiteClient
+    from vibe3.environment import SessionRegistryService
+    from vibe3.orchestra import FlowManagerProtocol
 
 
 def select_ready_issues_from_collected_issues(
